refactor : 변환기 클래스 추가 및 전체 코드 수정
- Converter 클래스 생성
- 오류 터트리는 위치 수정 Caclculator -> InputValidator
Close
- 다시 음수 예외 처리 내용 작성하기
- test code 다시 작성하기
- 정규식 다시 점검하기

+++ src/main/java/calculator/Controller.java
@@ -10,7 +10,7 @@ public class Controller {
         promptForInputString();
         String input = Console.readLine();
 
-        double add = calculator.add(input);
+        int add = calculator.add(input);
         printResult(add);
     }
 
@@ -18,7 +18,7 @@ public class Controller {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
     }
 
-    private void printResult(double add) {
+    private void printResult(int add) {
         System.out.println("결과 : " + add);
     }
 }
